ai_to_pixel_art/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pixel_art_converter"
version = "0.1.0"
authors = [
  { name="Example Author", email="author@example.com" },
]
description = "Batch convert AI-generated images into authentic pixel art."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Pillow>=9.0.0",
]

[project.scripts]
pixelart = "pixel_art_converter.cli:main"
